requirements.txt
aiogram~=3.21.0
SQLAlchemy~=2.0.42
httpx~=0.28.1
fastapi~=0.116.1
Pyrogram~=2.0.106
dotenv~=0.9.9
python-dotenv~=1.1.1
python-jose~=3.5.0
uvicorn